requirements.txt
# for OpenAI Privacy Filter
huggingface_hub
numpy
packaging
torch
safetensors
tiktoken
# for RESTful API
fastapi
uvicorn[standard]
pydantic
python-dotenv